fix logv2 active file size counter, replace log with logv2
